archive: use correct epoch time.
epoch is 1970-01-01 00:00:00 UTC, not 1970-01-01 01:00:00 UTC.

This mistake was introduced in commit ef5a32c to fix the test suite. I guess
that it was a hack that workaround'ed an actual bug for a specific timezone
only. If that bug reappears, we'll have to fix that by making this module
timezone-aware, not by setting a wrong epoch value.